refactor(ui): further flatten `cluster-workflow-templates/` directory
- after previous flattening, it was one `index.ts` file and then a `components/` directory with everything else
  - just flatten those into one, simplifying reading with less nesting

- also partially sort imports (similar to isort) by putting a newline between external deps and internal deps
  - TODO: add an `eslint` plugin for this now that we're on `eslint` / off `tslint`
