+++ ui/src/app/cluster-workflow-templates/cluster-workflow-template-container.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import {Route, RouteComponentProps, Switch} from 'react-router';
+
 import {ClusterWorkflowTemplateDetails} from './cluster-workflow-template-details';
 import {ClusterWorkflowTemplateList} from './cluster-workflow-template-list';
