+++ ui/src/app/cluster-workflow-templates/cluster-workflow-template-creator.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react';
 import {useState} from 'react';
 
-import {ClusterWorkflowTemplate} from '../../../models';
-import {Button} from '../../shared/components/button';
-import {ErrorNotice} from '../../shared/components/error-notice';
-import {ExampleManifests} from '../../shared/components/example-manifests';
-import {UploadButton} from '../../shared/components/upload-button';
-import {exampleClusterWorkflowTemplate} from '../../shared/examples';
-import {services} from '../../shared/services';
+import {ClusterWorkflowTemplate} from '../../models';
+import {Button} from '../shared/components/button';
+import {ErrorNotice} from '../shared/components/error-notice';
+import {ExampleManifests} from '../shared/components/example-manifests';
+import {UploadButton} from '../shared/components/upload-button';
+import {exampleClusterWorkflowTemplate} from '../shared/examples';
+import {services} from '../shared/services';
 import {ClusterWorkflowTemplateEditor} from './cluster-workflow-template-editor';
 
 export function ClusterWorkflowTemplateCreator({onCreate}: {onCreate: (workflow: ClusterWorkflowTemplate) => void}) {
