+++ ui/src/app/cluster-workflow-templates/cluster-workflow-template-details.tsx
@@ -3,17 +3,18 @@ import {SlidingPanel} from 'argo-ui/src/index';
 import * as React from 'react';
 import {useContext, useEffect, useState} from 'react';
 import {RouteComponentProps} from 'react-router';
-import {ClusterWorkflowTemplate} from '../../../models';
-import {uiUrl} from '../../shared/base';
-import {ErrorNotice} from '../../shared/components/error-notice';
-import {Loading} from '../../shared/components/loading';
-import {useCollectEvent} from '../../shared/components/use-collect-event';
-import {Context} from '../../shared/context';
-import {historyUrl} from '../../shared/history';
-import {services} from '../../shared/services';
-import {useQueryParams} from '../../shared/use-query-params';
-import {Utils} from '../../shared/utils';
-import {SubmitWorkflowPanel} from '../../workflows/components/submit-workflow-panel';
+
+import {ClusterWorkflowTemplate} from '../../models';
+import {uiUrl} from '../shared/base';
+import {ErrorNotice} from '../shared/components/error-notice';
+import {Loading} from '../shared/components/loading';
+import {useCollectEvent} from '../shared/components/use-collect-event';
+import {Context} from '../shared/context';
+import {historyUrl} from '../shared/history';
+import {services} from '../shared/services';
+import {useQueryParams} from '../shared/use-query-params';
+import {Utils} from '../shared/utils';
+import {SubmitWorkflowPanel} from '../workflows/components/submit-workflow-panel';
 import {ClusterWorkflowTemplateEditor} from './cluster-workflow-template-editor';
 
 import '../../../workflows/components/workflow-details/workflow-details.scss';
