+++ ui/src/app/cluster-workflow-templates/cluster-workflow-template-list.scss
